requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
groq==0.7.0
voyageai==0.2.1
numpy==1.26.4   # opcional mas geralmente necessário
pydantic==2.7.0
httpx==0.24.1